requirements_pi_docker.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
inference-sdk>=0.9.0
python-engineio>=4.8.0
python-socketio>=5.10.0
numpy>=1.24.0,<2.0.0
